Misc cleanup due to stricter type checking
Fixing various issues being checked by the compiler:
 - Changing generics to match callers
 - Calling invoke on the correct class
 - Fixing some ambiguous casts.
 - Fixing some method visibility issues.
 - Removing extra comma after parameters

+++ gemfire-core/src/test/java/com/gemstone/gemfire/management/UniversalMembershipListenerAdapterDUnitTest.java
@@ -1861,7 +1861,7 @@ public class UniversalMembershipListenerAdapterDUnitTest extends ClientServerTes
     
     // gather details for later creation of pool...
     assertEquals(ports[0],
-                 vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getTestServerEventsInLonerClient_port()));
+                 (int) vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getTestServerEventsInLonerClient_port()));
     String serverMemberId = (String) vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getMemberId());
     DistributedMember serverMember = (DistributedMember) vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getDistributedMember());
 
@@ -2029,7 +2029,7 @@ public class UniversalMembershipListenerAdapterDUnitTest extends ClientServerTes
     
     // gather details for later creation of pool...
     assertEquals(ports[0],
-                 vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getTestServerEventsInLonerClient_port()));
+        (int) vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getTestServerEventsInLonerClient_port()));
     serverMemberId = (String) vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getMemberId());
     serverMember = (DistributedMember) vm0.invoke(() -> UniversalMembershipListenerAdapterDUnitTest.getDistributedMember());
